requirements.txt
# SEC Financial QA System - Core Requirements
torch>=2.0.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
numpy>=1.24.0
pandas>=2.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
sec-edgar-downloader>=5.0.0
python-dotenv>=1.0.0

# Development dependencies (optional)
pytest>=7.4.0
black>=23.0.0
